исправление критериев
