+++ src/constants.js
@@ -1,14 +1,4 @@
-export const TASK_NUM = 22;
-export const MAX_TAGS_NUMBER = 3;
-export const TASK_LOAD_NUMB = 8;
-export const TAGS = [`homework`, `theory`, `practice`, `intensive`, `keks`];
-export const DESCRIPTIONS = [
-  `Изучить теорию`,
-  `Сделать домашку`,
-  `Пройти интенсив на соточку`,
-];
-
-export const TASK_DAY = {
+export const TaskDay = {
   MO: `mo`,
   TU: `tu`,
   WE: `we`,
@@ -17,14 +7,6 @@ export const TASK_DAY = {
   SA: `sa`,
   SU: `su`,
 };
-
-export const COLORS = [
-  `black`,
-  `yellow`,
-  `blue`,
-  `green`,
-  `pink`,
-];
 
 export const Time = {
   WEEK: 7 * 24 * 60 * 60 * 1000,
