+++ src/mock.js
@@ -1,4 +1,23 @@
-import {TAGS, MAX_TAGS_NUMBER, DESCRIPTIONS, TASK_DAY, Time, COLORS, TASK_NUM} from './constants';
+import {TaskDay, Time} from './constants';
+
+const TASK_NUM = 22;
+const MAX_TAGS_NUMBER = 3;
+const TAGS = [`homework`, `theory`, `practice`, `intensive`, `keks`];
+const DESCRIPTIONS = [
+  `Изучить теорию`,
+  `Сделать домашку`,
+  `Пройти интенсив на соточку`,
+];
+
+export const TASK_LOAD_NUMB = 8;
+
+export const COLORS = [
+  `black`,
+  `yellow`,
+  `blue`,
+  `green`,
+  `pink`,
+];
 
 const getRandomBool = (chance = 0.5) =>
   Math.random() > chance;
@@ -9,7 +28,7 @@ const getRandomItem = (array) =>
 const getRandomNumber = (min = 0, max) =>
   Math.floor(Math.random() * (max - min + 1)) + min;
 
-const weekDays = Object.values(TASK_DAY);
+const weekDays = Object.values(TaskDay);
 
 const repeatDayReducer = (days, day) => {
   days[day] = getRandomBool(0.9);
